+++ src/Headers/index.browser.js
@@ -1,2 +1,0 @@
-/** @type {import('.').Headers} */
-export const { Headers } = window