+++ src/fetch_method/index.d.ts
@@ -1,0 +1,10 @@
+export declare function fetch_method__new(params:fetch_method___params_T):string
+export {
+	fetch_method__new as fetch_method_,
+	fetch_method__new as _fetch_method,
+	fetch_method__new as _method__fetch,
+}
+export declare interface fetch_method___params_T {
+	method:string
+}
+export declare type fetch_method_params__T = fetch_method___params_T