+++ src/fetch_method_/index.d.ts
@@ -1,9 +1,0 @@
-export declare function fetch_method_(params:fetch_method___params_T):string
-export {
-  fetch_method_ as _fetch_method,
-  fetch_method_ as _method__fetch,
-}
-export declare interface fetch_method___params_T {
-	method:string
-}
-export declare type fetch_method_params__T = fetch_method___params_T