+++ src/fetch_response_pair/index.browser.js
@@ -1,0 +1,5 @@
+import { fetch_response_pair__new2 } from './index.js'
+export const fetch_response_pair__new = fetch_response_pair__new2(window.fetch)
+export {
+	fetch_response_pair__new as fetch_response_pair_,
+}