+++ src/fetch_response_pair_/index.browser.js
@@ -1,3 +1,0 @@
-import { fetch_response_pair__ } from '../fetch_response_pair__/index.js'
-/** @type {typeof import('../index.d.ts').fetch_response_pair_)} */
-export const fetch_response_pair_ = fetch_response_pair__(window.fetch)