+++ src/fetch_response_pair_/index.d.ts
@@ -1,3 +1,0 @@
-export declare function fetch_response_pair_<Body>(
-	input:RequestInfo, init?:RequestInit
-):Promise<[Body, globalThis.Response]>