+++ src/fetch_response_pair__/index.d.ts
@@ -1,6 +1,0 @@
-import type { fetch_response_pair_ } from '../fetch_response_pair_'
-import type { _fetch } from '../_types'
-export declare function fetch_response_pair__(fetch:_fetch):typeof fetch_response_pair_
-export {
-	fetch_response_pair__ as fetch_response_pair_2
-}