+++ src/headers_/index.browser.js
@@ -1,9 +1,0 @@
-/**
- * @param {HeadersInit}init
- * @returns {HeadersInit|undefined}
- * @private
- */
-export function headers_(init) {
-	return new window.Headers(init)
-}
-export { headers_ as _headers, }